+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Added Markdown report output with `--format markdown`
+
 ## 0.1.0
 
 - Initial release readiness CLI